refactor(spec): split spec.rs into src/spec/ module
Purely mechanical extraction — zero logic changes. Splits the 2,117-line
src/spec.rs into src/spec/{mod,parse,validation,commands,tests}.rs while
preserving all public APIs (SpecAction, SpecFrontmatter, IndexEntry, run,
collect_index, render_index_markdown, all_module_names,
specs_for_changed_files, load_module_bundle) importable from crate::spec::.

Also updates specs/spec/spec.spec.md files: list to reference the new
submodule paths instead of the deleted src/spec.rs.
